refactor: Update metadata field in ExperimentConfig to use Field for default factory; improve code readability and consistency refactor: Adjust indentation in 1Q Pauli matrix definition for clarity refactor: Update pyproject.toml to include additional ruff-specific rules and ignore settings refactor: Reorder __all__ exports in various modules for consistency refactor: Simplify label extraction in state tomography and inspection summary classes refactor: Enhance list handling in experiment services and measurement service for clarity refactor: Improve Hamiltonian construction in QuantumSimulator for better readability refactor: Refactor hex_to_rgba function to use unpacking for clarity refactor: Clean up test_measurement_schedule_builder.py by removing unnecessary comments

+++ src/qubex/pulse/pulse_schedule.py
@@ -163,7 +163,7 @@ class PulseSchedule:
         #   We therefore compute the ratio and round it to the nearest integer.
         #   This assumes that duration is always intended to be an integer multiple of
         #   Waveform.SAMPLING_PERIOD within normal floating point tolerance.
-        return int(round(self.duration / Waveform.SAMPLING_PERIOD))
+        return round(self.duration / Waveform.SAMPLING_PERIOD)
 
     @property
     def duration(self) -> float:
